{"schema":"decodex/commit/1","summary":"Split Swift operator run strip boundaries","authority":"manual"}
