+++ apps/decodex-app/Sources/DecodexApp/OperatorLaneReadoutSupport.swift
@@ -1,0 +1,182 @@
+import Foundation
+import SwiftUI
+
+enum OperatorLaneReadoutLayout {
+	static let titleWidth: CGFloat = 92
+	static let columnSpacing: CGFloat = 7
+	static let itemRowSpacing: CGFloat = 2
+	static let progressTrackWidth: CGFloat = 84
+	static let totalSegmentSpacing: CGFloat = 13
+	static let totalValueLabelSpacing: CGFloat = 4
+	static let metricColumnCount = 4
+	static let lifecycleTableColumnSpacing: CGFloat = 18
+}
+
+enum OperatorLanePopoverStyle {
+	static let titleFont = PanelFont.laneTitle
+	static let projectFont = PanelFont.laneDetail
+	static let labelFont = PanelFont.lanePopoverLabel
+	static let valueFont = PanelFont.lanePopoverValue
+	static let metaFont = PanelFont.lanePopoverMeta
+	static let separatorFont = PanelFont.lanePopoverMeta
+
+	static func primaryText(_ colorScheme: ColorScheme) -> Color {
+		Color.primary.opacity(colorScheme == .dark ? 0.94 : 0.88)
+	}
+
+	static func secondaryText(_ colorScheme: ColorScheme) -> Color {
+		Color.secondary.opacity(colorScheme == .dark ? 0.92 : 0.86)
+	}
+
+	static func mutedText(_ colorScheme: ColorScheme) -> Color {
+		Color.secondary.opacity(colorScheme == .dark ? 0.78 : 0.68)
+	}
+
+	static func tableHeaderText(_ colorScheme: ColorScheme) -> Color {
+		Color.secondary.opacity(colorScheme == .dark ? 0.72 : 0.62)
+	}
+
+	static func separator(_ colorScheme: ColorScheme) -> Color {
+		Color.secondary.opacity(colorScheme == .dark ? 0.18 : 0.22)
+	}
+
+	static func progressTrack(_ colorScheme: ColorScheme) -> Color {
+		Color.secondary.opacity(colorScheme == .dark ? 0.2 : 0.22)
+	}
+
+	static func progressFill(_ colorScheme: ColorScheme) -> Color {
+		PanelPalette.routeAccent(colorScheme).opacity(colorScheme == .dark ? 0.86 : 0.76)
+	}
+}
+
+struct OperatorLaneReadoutItem: Identifiable {
+	let label: String?
+	let value: String
+
+	init(label: String?, value: String) {
+		self.label = label
+		self.value = value
+	}
+
+	var id: String {
+		"\(label ?? "value")-\(value)"
+	}
+
+	var displayValue: String {
+		if value.hasSuffix(" tok") {
+			return String(value.dropLast(4))
+		}
+
+		return value
+	}
+
+	var accessibilityText: String {
+		if let label {
+			return "\(label) \(displayValue)"
+		}
+
+		return displayValue
+	}
+
+	var summaryRuns: [OperatorLaneReadoutTextRun] {
+		switch label?.lowercased() {
+		case "wall":
+			return [.meta("wall "), .value(displayValue)]
+		case "project":
+			return [.meta(displayValue)]
+		case "attempts":
+			return [.value(displayValue), .meta(displayValue == "1" ? " attempt" : " attempts")]
+		case "captured":
+			return [.value(displayValue), .meta(" captured")]
+		case "missing":
+			return [.value(displayValue), .meta(" missing")]
+		case "events":
+			return [.value(displayValue), .meta(" events")]
+		case "child events":
+			return [.value(displayValue), .meta(" child events")]
+		case "input", "input tokens":
+			return [.value(displayValue), .meta(" input")]
+		case "output", "output tokens":
+			return [.value(displayValue), .meta(" output")]
+		case "current":
+			return [.value(displayValue), .meta(" current")]
+		case "current window", "current window tokens":
+			return [.value(displayValue), .meta(" current window")]
+		case "peak":
+			return [.value(displayValue), .meta(" peak")]
+		case "peak window", "peak window tokens":
+			return [.value(displayValue), .meta(" peak window")]
+		case "tools", "tool calls":
+			return [.value(displayValue), .meta(" tools")]
+		case "output bytes":
+			return [.value(displayValue), .meta(" output")]
+		case "max output", "max tool output", "largest output":
+			return [.value(displayValue), .meta(" max output")]
+		case "largest tool", "source":
+			return [.value(displayValue), .meta(" source")]
+		default:
+			if let label {
+				return [.meta("\(label) "), .value(displayValue)]
+			}
+			return [.value(displayValue)]
+		}
+	}
+
+	func matchesLabel(_ expected: String) -> Bool {
+		label?.caseInsensitiveCompare(expected) == .orderedSame
+	}
+}
+
+enum OperatorLaneReadoutTextRole {
+	case meta
+	case value
+}
+
+struct OperatorLaneReadoutTextRun {
+	let text: String
+	let role: OperatorLaneReadoutTextRole
+
+	static func meta(_ text: String) -> OperatorLaneReadoutTextRun {
+		OperatorLaneReadoutTextRun(text: text, role: .meta)
+	}
+
+	static func value(_ text: String) -> OperatorLaneReadoutTextRun {
+		OperatorLaneReadoutTextRun(text: text, role: .value)
+	}
+}
+
+struct OperatorTotalMetric: Identifiable {
+	let title: String
+	let items: [OperatorLaneReadoutItem]
+
+	var id: String {
+		title
+	}
+
+	var accessibilityText: String {
+		let itemText = items.map(\.accessibilityText).joined(separator: ", ")
+
+		return itemText.isEmpty ? title : "\(title), \(itemText)"
+	}
+}
+
+func rawPanelToken(_ value: String) -> String {
+	value.trimmingCharacters(in: .whitespacesAndNewlines)
+}
+
+struct OperatorLaneReadoutWidthKey: PreferenceKey {
+	static let defaultValue: CGFloat = 0
+
+	static func reduce(value: inout CGFloat, nextValue: () -> CGFloat) {
+		value = max(value, nextValue())
+	}
+}
+
+struct OperatorLaneReadoutWidthReader: View {
+	var body: some View {
+		GeometryReader { proxy in
+			Color.clear
+				.preference(key: OperatorLaneReadoutWidthKey.self, value: proxy.size.width)
+		}
+	}
+}
